brain-tumor-backend/requirements.txt
flask
tensorflow
flask-cors
pillow